+++ Types/ModificationPriceInfo.cs
@@ -3,8 +3,8 @@ namespace Automatization.Types
     public class ModificationPriceInfo
     {
         public string Modification { get; set; } = string.Empty;
-        public int BasePrice { get; set; }
-        public int DiscountAmount { get; set; }
-        public int FinalPrice { get; set; }
+        public long BasePrice { get; set; }
+        public long DiscountAmount { get; set; }
+        public long FinalPrice { get; set; }
     }
 }
